requirements.txt
certifi==2024.8.30
numpy==2.1.3
pandas==2.2.3
Flask>=2.3.2
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
weatherapipython @ git+https://github.com/weatherapicom/python.git@05344f2d04cbf0b9513cb69b71f56a66a348e1be
python-dotenv
